@@ -26,9 +26,9 @@
     "lodash": "^4.17.21",
     "notistack": "^2.0.2",
     "pako": "^2.0.4",
-    "react": "^17.0.2",
+    "react": "^18.0.0-rc.0",
     "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0-rc.0",
     "react-i18next": "^11.11.0",
     "react-redux": "^7.2.0",
     "react-scripts": "^4.0.3",
